# docs/api/api-spec/openapi-bundled.yaml
openapi: 3.0.3
info:
  title: Event Booking Platform API
  version: 0.1.0
  description: |
    API contract for a Spring Boot event booking backend. The MvP supports event browsing, seat availability, seat reservation, and booking confirmation
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security: []
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register an account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Successfully Register
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid Register Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Log in
      operationId: LogIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: Successfully log in
          content:
            application/jon:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid Login Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: SEAT_NOT_AVAILABLE
        message:
          type: string
          example: This seat is no longer available
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - fullName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        fullName:
          type: string
    RegisterResponse:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        email:
          type: string
          format: email
        fullName:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 3600